feat: improved error handling for secrets
